livekit-voice-agent/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
docstring-parser==0.17.0
eval-type-backport==0.3.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
google-api-core==2.30.1
google-auth==2.49.1
google-cloud-speech==2.38.0
google-cloud-texttospeech==2.35.0
google-genai==1.69.0
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.8.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.13.0
livekit==1.1.3
livekit-agents==1.5.1
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-google==1.5.1
livekit-plugins-murf==1.5.1
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.5.1
livekit-plugins-turn-detector==1.5.1
livekit-protocol==1.1.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
onnxruntime==1.24.4
openai==2.30.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
prometheus-client==0.24.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.20.0
pyjwt==2.12.1
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
zipp==3.23.0
